requirements.txt
streamlit==1.36.0
PyPDF2==3.0.1
langchain==0.2.5
langchain-google-genai==1.0.7
google-generativeai==0.7.0
faiss-cpu==1.8.0.post1
groq==0.9.0
python-dotenv==1.0.1
huggingface-hub==0.23.4
langchain-community
sentence-transformers
